+++ mi/dataset/driver/ctdav_n/auv/ctdav_n_auv_recovered_driver.py
@@ -9,8 +9,6 @@
 Release notes:
 
 Initial Release
-
-THIS FILE IS DEPRECATED, REPLACED BY mi/dataset/driver/ctdav_n/auv/ctdav_n_auv_driver.py
 """
 
 from mi.dataset.dataset_driver import SimpleDatasetDriver
@@ -22,7 +20,7 @@ from deprecated import deprecated
 @version("15.6.1")
 def parse(unused, source_file_path, particle_data_handler):
     """
-    This is the method called by Uframe
+    This is the method called by uFrame
     :param unused
     :param source_file_path This is the full path and filename of the file to be parsed
     :param particle_data_handler Java Object to consume the output of the parser
@@ -31,7 +29,7 @@ def parse(unused, source_file_path, particle_data_handler):
 
     with open(source_file_path, 'rU') as stream_handle:
 
-        # create and instance of the concrete driver class defined below
+        # Create an instance of the concrete driver class defined below.
         driver = CtdavNAuvRecoveredDriver(unused, stream_handle, particle_data_handler)
         driver.processFileStream()
 
@@ -41,8 +39,7 @@ def parse(unused, source_file_path, particle_data_handler):
 @deprecated
 class CtdavNAuvRecoveredDriver(SimpleDatasetDriver):
     """
-    Derived adcpa_n_auv driver class
-    All this needs to do is create a concrete _build_parser method
+    Create a concrete _build_parser method for the adcpa_n_auv driver.
     """
 
     def _build_parser(self, stream_handle):