+++ mi/dataset/driver/ctdav_nbosi/auv/ctdav_nbosi_auv_driver.py
@@ -19,7 +19,7 @@ from mi.core.versioning import version
 @version("0.1.0")
 def parse(unused, source_file_path, particle_data_handler):
     """
-    This is the method called by Uframe
+    This is the method called by uFrame
     :param unused
     :param source_file_path This is the full path and filename of the file to be parsed
     :param particle_data_handler Java Object to consume the output of the parser
@@ -37,8 +37,7 @@ def parse(unused, source_file_path, particle_data_handler):
 
 class CtdavNboisAuvDriver(SimpleDatasetDriver):
     """
-    Derived adcpa_n_auv driver class
-    All this needs to do is create a concrete _build_parser method
+    Create a concrete _build_parser method for the adcpa_n_auv driver.
     """
 
     def _build_parser(self, stream_handle):